modelserver/requirements.txt
fastapi
pika  
uvicorn
python-multipart
python-dotenv
pillow
av
asyncio
numpy>=1.21.6,<1.28.0
scikit-learn==1.3.2
scipy==1.11.4
python-dotenv
torch
torchvision
torchaudio
pydantic==1.9.0
tqdm
matplotlib
transformers>=4.0.0

